agents_live-main/requirements_files.txt
# ============================================================================
# requirements.txt - Core Dependencies (All Engines)
# ============================================================================

# Async Runtime
asyncio==3.4.3
aiofiles==23.2.1

# NATS Messaging
nats-py==2.6.0

# Database
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.12.1

# Redis Cache
redis==5.0.1

# HTTP Client
aiohttp==3.9.1
httpx==0.25.2

# System Monitoring
psutil==5.9.6

# Data Processing
numpy==1.26.2
pandas==2.1.4

# Logging
structlog==23.2.0
python-json-logger==2.0.7

# Configuration
pydantic==2.5.2
python-dotenv==1.0.0
pyyaml==6.0.1

# Observability
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-asyncpg==0.42b0
opentelemetry-instrumentation-redis==0.42b0
prometheus-client==0.19.0

# Security
cryptography==41.0.7
pyjwt==2.8.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7

# ============================================================================
# requirements-optimizing.txt - Optimizing Engine Specific
# ============================================================================

# Scientific Computing
scipy==1.11.4

# Statistics
statistics==1.0.3.5

# Advanced Math
sympy==1.12

# ============================================================================
# requirements-performance.txt - Performance Engine Specific
# ============================================================================

# Machine Learning for Anomaly Detection
scikit-learn==1.3.2

# Time Series Analysis
statsmodels==0.14.1

# Performance Profiling
py-spy==0.3.14
memory-profiler==0.61.0

# ============================================================================
# requirements-analyzer.txt - Analyzer Engine Specific
# ============================================================================

# Code Parsing
ast-grep==0.1.0

# Security Analysis
bandit==1.7.5
safety==2.3.5

# Code Quality
pylint==3.0.3
flake8==6.1.0
mypy==1.7.1
black==23.12.0
isort==5.13.2

# Complexity Analysis
radon==6.0.1
mccabe==0.7.0

# Machine Learning
scikit-learn==1.3.2
joblib==1.3.2

# NLP for code understanding
spacy==3.7.2

# ============================================================================
# requirements-dev.txt - Development Dependencies
# ============================================================================

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-timeout==2.2.0
hypothesis==6.92.1

# Test Utilities
faker==20.1.0
factory-boy==3.3.0

# Debugging
ipdb==0.13.13
ipython==8.18.1

# Code Quality
pre-commit==3.6.0
ruff==0.1.8

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0
sphinx-autodoc-typehints==1.25.2

# Type Checking
mypy==1.7.1
types-redis==4.6.0.11
types-requests==2.31.0.10

# Performance Testing
locust==2.18.3

# ============================================================================
# requirements-prod.txt - Production Only
# ============================================================================

# Production WSGI Server (if needed for HTTP endpoints)
gunicorn==21.2.0

# APM
elastic-apm==6.19.0
sentry-sdk==1.39.1

# Health Checks
py-healthcheck==1.10.1

# Rate Limiting
aiolimiter==1.1.0

# Retry Logic
tenacity==8.2.3

# Circuit Breaker
pybreaker==1.0.1

# Connection Pooling
aiomysql==0.2.0  # if MySQL needed
aiopg==1.4.0     # additional PostgreSQL support

# ============================================================================
# constraints.txt - Version Constraints for Security
# ============================================================================

# Pin critical security dependencies
certifi>=2023.7.22
urllib3>=2.0.7
requests>=2.31.0

# Ensure compatible versions
python-dateutil>=2.8.2,<3.0.0
pytz>=2023.3

# Database drivers
asyncpg>=0.29.0,<1.0.0
psycopg2-binary>=2.9.9

# Redis
redis>=5.0.0,<6.0.0

# Cryptography
cryptography>=41.0.0,<42.0.0
